Move app component to containers
