+++ src/__tests__/App.test.js
@@ -1,5 +1,5 @@
 import { create } from 'react-test-renderer';
-import App from '../components/App';
+import App from '../containers/App';
 
 test('renders correctly', () => {
   const tree = create(<App />).toJSON();
