+++ src/index.js
@@ -4,7 +4,7 @@ import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
 import rootReducer from './reducers';
-import App from './components/App';
+import App from './containers/App';
 import './index.css';
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
